Cast division factors to doubles

+++ app_radio_module/src/transmitter.c
@@ -102,8 +102,8 @@ static void gnss_data_cb(const struct device *dev, const struct gnss_data *data)
     packet.payload_len = sizeof(l_gnss_data_t);
 
     l_gnss_data_t gnss_data = {0};
-    gnss_data.latitude = (double)data->nav_data.latitude / L_GNSS_LATITUDE_DIVISION_FACTOR;
-    gnss_data.longitude = (double)data->nav_data.longitude / L_GNSS_LONGITUDE_DIVISION_FACTOR;
+    gnss_data.latitude = (double)data->nav_data.latitude / (double)L_GNSS_LATITUDE_DIVISION_FACTOR;
+    gnss_data.longitude = (double)data->nav_data.longitude / (double)L_GNSS_LONGITUDE_DIVISION_FACTOR;
     gnss_data.altitude = (float)data->nav_data.altitude / L_GNSS_ALTITUDE_DIVISION_FACTOR;
 
     memcpy(packet.payload, &gnss_data, sizeof(l_gnss_data_t));
